python-claude-flow/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-flow"
version = "2.0.0-alpha.90"
description = "Enterprise-grade AI agent orchestration with ruv-swarm integration (Alpha Release)"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "rUv", email = "info@ruv.net"}
]
keywords = [
    "claude", "ai", "agent", "orchestration", "mcp", "workflow", 
    "automation", "swarm", "ruv-swarm", "github", "docker", 
    "enterprise", "coordination", "multi-agent", "neural-networks", 
    "cli", "tools", "alpha"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "asyncio-mqtt>=0.16.0",
    "websockets>=11.0.0",
    "aiohttp>=3.8.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "aiosqlite>=0.19.0",
    "redis>=4.5.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "langchain>=0.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "psutil>=5.9.0",
    "watchdog>=3.0.0",
    "colorama>=0.4.6",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
ui = [
    "blessed>=1.20.0",
    "prompt-toolkit>=3.0.0",
]

[project.scripts]
claude-flow = "claude_flow.cli.main:main"

[project.urls]
Homepage = "https://github.com/ruvnet/claude-code-flow"
Repository = "https://github.com/ruvnet/claude-code-flow"
Documentation = "https://github.com/ruvnet/claude-code-flow#readme"
Issues = "https://github.com/ruvnet/claude-code-flow/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
